changed code, now much cleaner and added better support for reading materials

+++ Sourcefiles/Camera.h
@@ -14,7 +14,7 @@ class Camera
 	public:
 		Camera(Vector3& position,Vector3& lookAt,Vector3& up,float fov = 60.0f);
 
-		inline Ray GetRayForPixel(int x,int y,float ratio)const;
+		inline Ray GetRayForPixel(float x,float y,float ratio)const;
 
 		const Vector3& GetOrigin()const {return m_position;}
 
@@ -24,16 +24,11 @@ class Camera
 		float m_fov,m_distance;
 };
 
-Ray Camera::GetRayForPixel(int x,int y,float ratio)const
+Ray Camera::GetRayForPixel(float x,float y,float ratio)const
 {
-	//assert(x >= 0 && y >= 0 && "x or y are less than 0!");
-	//assert(x < imgSize && y < imgSize && "x or y are bigger than image size!");
 
-	Vector3 pixelDir = m_distance * m_lookVector + (0.5f - (float)y * ratio) * m_upVector +
-		             ((float)x * ratio - 0.5f) * m_rightVector;
-
-	//Vector3 pixelDir = Vector3((float)x,(float)y,0.0f) - m_position;
-	//pixelDir = Normalize(pixelDir);
+	Vector3 pixelDir = m_distance * m_lookVector + (0.5f - y * ratio) * m_upVector +
+		             (x * ratio - 0.5f) * m_rightVector;
 
 
 	return Ray(pixelDir,m_position);
